Refactor 2D DDDM code into modular package
