+++ run_example.py
@@ -1,15 +1,23 @@
+"""Minimal example to train & sample on 2D GMM, matching Section 6.1."""
+import argparse
+import json
+import os
 
-# Minimal example to train & sample on 2D GMM, matching Section 6.1.
-# Usage:
-#   python run_example.py --epochs 1000 --beta 0.1 --lam 1.0 --m 8 --steps 20
-import argparse, os, json
 import torch
 
-from dddm_2d import TrainConfig, train_dddm, sample_dddm, sample_gmm, rbf_mmd2, save_scatter
+from dddm import (
+    TrainConfig,
+    train_dddm,
+    sample_dddm,
+    sample_gmm,
+    rbf_mmd2,
+    save_scatter,
+)
 
-def main():
+
+def main() -> None:
     p = argparse.ArgumentParser()
-    p.add_argument("--epochs", type=int, default=100000)
+    p.add_argument("--epochs", type=int, default=1000)
     p.add_argument("--batch", type=int, default=512)
     p.add_argument("--beta", type=float, default=0.1)
     p.add_argument("--lam", type=float, default=1.0)
@@ -21,8 +29,16 @@ def main():
     p.add_argument("--out", type=str, default="./out")
     args = p.parse_args()
 
-    cfg = TrainConfig(beta=args.beta, lam=args.lam, m=args.m, w_bias=args.w_bias,
-                      epochs=args.epochs, batch=args.batch, device=args.device, seed=args.seed)
+    cfg = TrainConfig(
+        beta=args.beta,
+        lam=args.lam,
+        m=args.m,
+        w_bias=args.w_bias,
+        epochs=args.epochs,
+        batch=args.batch,
+        device=args.device,
+        seed=args.seed,
+    )
     os.makedirs(args.out, exist_ok=True)
 
     model = train_dddm(cfg, outdir=args.out)
@@ -38,6 +54,7 @@ def main():
         json.dump({"mmd2_rbf_sigma1": mmd2}, f, indent=2)
     print(f"MMD^2 (rbf σ=1) = {mmd2:.4f}")
     print(f"Saved samples and metrics in {args.out}")
+
 
 if __name__ == "__main__":
     main()
